Install Alpine.js as a dependency
Alpine.js is used at runtime.  Alpine.js is not just a build-time tool.
It runs in the browser and it must be included in the final JavaScript
bundle that gets sent to the user.

So even though Alpine is small (~10KB), it must be present in production
builds.

$ sail npm install alpinejs

added 6 packages, and audited 84 packages in 2s