Use nesting components for base layout
Elevate the base layout from a flat view and rebuild it with a nesting
of multilevel components for re-usability, encapsulation, consistency,
and easy maintenance.  This approach makes Blade views very modular.
Separation of concerns helps the components be isolated.  Each piece of
UI is focused on its own and contributes individually to the whole Blade
views.