Use wire:model.defer for updates
Livewire uses morphing to update parts of the DOM after an action like
typing in an input and tabbing away.

The wire:target tells Livewire to re-render the block whenever the $name
property changes.

But Livewire doesn’t know why or how $name changed unless the change was
caused by something with a wire: directive.  Use wire:model.defer to
update the back-end $name property when input loses focus.

Putting both the input and the output ({{ $name }}) inside the same
wire:target block makes it crystal clear to Livewire what part of the
page needs to be updated when $name changes.  The connection between the
input and the target isn't direct and it is Livewire's diffing engine
that decides what gets updated.  But wrapping them together makes it
predictable and easier to debug.

Use wire:change="$refresh" to force Livewire to re-render the component
and update any wire:target sections.