Remove wire:target="name"
Livewire automatically tracks {{ $name }} without needing wire:target.
We don't need wire:target="name" if we use {{ $name }}.

Livewire 3 parses the Blade template, builds a map of which variables
are used where in the DOM, and uses that to determine which parts of the
page need updating after an interaction.