Add an argument to a Livewire method
In Livewire 3, defining a method with required or optional parameters,
Livewire will silently ignore them if not provided from the front end as
long as the method can be called without arguments, i.e., default values
are defined.  PHP allows calling the method without any arguments when a
default is provided.

Livewire internally does something like this: $this->resetName(), and
PHP resolves it automatically to $this->resetName('The plan is ready.').
So everything works fine, and no error will ever be thrown.

If we remove the default value, PHP will throw an error because $name
has no default and Livewire doesn't pass anything.

The current code is valid.  We're using a default value.  Livewire
doesn’t need to pass anything.  PHP fills in the gap automatically.

One of the elegant aspects of Livewire is how it allows you to write
expressive, reusable methods, all while remaining flexible where it
counts.

Add the missing method type of void.