Use wire:submit for action in form
Build the form to submit with the action by wire:submit.prevent to be
executed by the button.  It is important to use the button type of
submit with prevent specified in wire:submit since we do not intend to
save anything in the database.  Using <button> element for action is
semantically correct and good for accessibility.

Use consistent styling for both desktop and mobile and make the view
look appealing.