Use $set() in lieu of custom method
The $set() is a magic method provided by Livewire to help set public
component properties directly from the front end.

The use $set method is quick and simple, not requiring to define a new
method and reducing boilerplate, especially useful for small updates.
It is also great for prototyping and speeds up development, but only
works with public properties.

But it is less reusable	and can’t encapsulate logic or side effects.
Not ideal for complex actions when we can't validate, log, or chain
behavior easily.

We do not need to call $refresh after $set() unless we are updating
computed data that does not trigger reactivity automatically.

Livewire 3 handles the DOM updates automatically either through $set()
or a custom method.