Add mount method to Livewire component class
The mount method is a life cycle hook that runs once when the component
is first loaded.  It’s used to initialize properties, fetch data from
the database or API and set default values based on request parameters.

When the page loads and Livewire initializes this component, the mount
method runs before rendering.  It sets $this->name to 'Mounting the
plan.' instead of the default "We're making a plan.".

Any public property initialized at the class level (like $name) will be
overridden if we set it again inside mount().