Build todo-list Livewire component
Use artisan tool to generate component scaffolding.

$ sail artisan make:livewire -- todo-list
 COMPONENT CREATED  🤙

CLASS: app/Livewire/TodoList.php
VIEW:  resources/views/livewire/todo-list.blade.php

Build the class with the array of the todos list and use public access
modifier to easily pass it to Livewire view component.

Use the passed-in array of todos list and build the Livewire view for an
easy and attractive display.