Build remove method for todos list
Use array_pop() PHP built-in to build the remove method with no need to
check if the array is empty for simplicity.

Use type="button" to execute the button for Livewire method so that the
browser knows it is not for form submit.

The only downside is it removes the last item rather than the designated
one, yet the simplicity of the design warrants this approach.  In most
cases we do want to remove the last item only when we make a mistake.